Make lists and struct expandable
